Cleanups to remove or mitigate risk of panics.
- Documents what users can expect panic-wise.
- Minimizes visibility of code that deals with internal indexes directly.
- Uses checked logic for accessing internal array, or documents why direct indexing is safe.
