+++ src/iterators/log.rs
@@ -39,7 +39,7 @@ impl<'a, T: 'a + Counter> Iter<'a, T> {
 impl<'a, T: 'a + Counter> PickyIterator<T> for Iter<'a, T> {
     fn pick(&mut self, index: usize, _: u64) -> bool {
         let val = self.hist.value_for(index);
-        if val >= self.current_step_lowest_value_reporting_level || index == self.hist.last() {
+        if val >= self.current_step_lowest_value_reporting_level || index == self.hist.last_index() {
             // implies log_base must be > 1.0
             self.next_value_reporting_level *= self.log_base;
             // won't underflow since next_value_reporting_level starts > 0 and only grows
